intent_service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
